pattern utils
